requirements.txt
# PDF Processing
pymupdf>=1.24.0

# Embeddings & LLM (NVIDIA API via OpenAI client)
openai>=1.0.0

# Vector Database
chromadb>=0.4.0

# Data Models & Validation
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# CLI Output
rich>=13.0.0

# Testing
pytest>=8.0.0
